+++ lib/solid_objects/actor_definition.rb
@@ -13,6 +13,8 @@ module SolidObjects
     # @rbs @state_migrations: Array[StateMigration]
     # @rbs @activation_hooks: Array[Proc]
     # @rbs @deactivation_hooks: Array[Proc]
+    # @rbs @attribute_queries: Hash[Symbol, bool]
+    # @rbs @method_messages: Hash[Symbol, bool]
 
     attr_reader :state_definition,
       :messages,
@@ -33,6 +35,24 @@ module SolidObjects
       @state_migrations = []
       @activation_hooks = []
       @deactivation_hooks = []
+      @attribute_queries = {}
+      @method_messages = {}
+    end
+
+    # @rbs (Symbol | String, default: untyped) -> StateDefinition::Attribute
+    def add_attribute(name, default:)
+      attribute_name = name.to_sym
+      if messages.key?(attribute_name) || queries.key?(attribute_name)
+        raise InvalidActor, "#{attribute_name.inspect} is already defined"
+      end
+
+      state_definition.attribute(attribute_name, default:).tap do
+        queries[attribute_name] = Handler.new(
+          name: attribute_name,
+          block: -> { state.fetch(attribute_name) }
+        )
+        attribute_queries[attribute_name] = true
+      end
     end
 
     # @rbs (Symbol | String, Proc) -> Handler
@@ -42,6 +62,13 @@ module SolidObjects
 
     # @rbs (Symbol | String, Proc) -> Handler
     def add_query(name, block)
+      query_name = name.to_sym
+      if attribute_queries.delete(query_name)
+        return Handler.new(name: query_name, block:).tap do |handler|
+          queries[query_name] = handler
+        end
+      end
+
       add_handler(queries, name, block)
     end
 
@@ -54,6 +81,19 @@ module SolidObjects
       Handler.new(name: observable_name, block: observable_block).tap do |handler|
         observables[observable_name] = handler
       end
+    end
+
+    # @rbs (Class) -> ActorDefinition
+    def synchronize_instance_messages(actor_class)
+      names = actor_message_method_names(actor_class)
+
+      (method_messages.keys - names).each do |name|
+        messages.delete(name)
+        method_messages.delete(name)
+      end
+
+      names.each { |name| add_method_message(name) unless method_messages.key?(name) }
+      self
     end
 
     # @rbs (Integer) -> Integer
@@ -111,10 +151,55 @@ module SolidObjects
         copy.instance_variable_set(:@state_migrations, state_migrations.dup)
         copy.instance_variable_set(:@activation_hooks, activation_hooks.dup)
         copy.instance_variable_set(:@deactivation_hooks, deactivation_hooks.dup)
+        copy.instance_variable_set(:@attribute_queries, attribute_queries.dup)
+        copy.instance_variable_set(:@method_messages, method_messages.dup)
       end
     end
 
     private
+
+    attr_reader :attribute_queries, :method_messages
+
+    # @rbs (Symbol) -> Handler
+    def add_method_message(name)
+      if messages.key?(name) || queries.key?(name)
+        raise InvalidActor, "#{name.inspect} is already defined"
+      end
+
+      Handler.new(
+        name:,
+        block: ->(**arguments) { public_send(name, **arguments) }
+      ).tap do |handler|
+        messages[name] = handler
+        method_messages[name] = true
+      end
+    end
+
+    # @rbs (Class) -> Array[Symbol]
+    def actor_message_method_names(actor_class)
+      actor_ancestors = actor_class.ancestors.take_while { |ancestor| ancestor != Actor }
+      attribute_method_names = actor_class.send(:generated_attribute_methods).keys
+      validate_attribute_methods!(attribute_method_names)
+
+      actor_class.public_instance_methods.filter_map do |name|
+        method_name = name.to_sym
+        next if attribute_method_names.include?(method_name)
+        next unless actor_ancestors.include?(actor_class.instance_method(method_name).owner)
+
+        method_name
+      end.uniq
+    end
+
+    # @rbs (Array[Symbol]) -> void
+    def validate_attribute_methods!(attribute_method_names)
+      state_definition.to_h.each_key do |name|
+        [ name, :"#{name}=" ].each do |method_name|
+          next if attribute_method_names.include?(method_name)
+
+          raise InvalidActor, "actor method #{method_name} overrides attribute #{name.inspect}"
+        end
+      end
+    end
 
     # @rbs (Hash[Symbol, Handler], Symbol | String, Proc) -> Handler
     def add_handler(collection, name, block)