+++ lib/solid_objects/reference.rb
@@ -52,9 +52,40 @@ module SolidObjects
       )
     end
 
+    # @rbs (Symbol, *untyped, **untyped) -> untyped
+    def method_missing(name, *arguments, **keywords)
+      return super if arguments.any? || block_given?
+      return tell(name, **keywords) if actor_message?(name)
+      return Serialization.readonly_copy(ask(name, **keywords)) if actor_query?(name)
+
+      super
+    end
+
+    # @rbs (Symbol, bool) -> bool
+    def respond_to_missing?(name, include_private = false)
+      actor_message?(name) || actor_query?(name) || super
+    end
+
     # @rbs () -> String
     def to_s
       "#{actor_type}(#{actor_id.inspect})"
+    end
+
+    private
+
+    # @rbs (Symbol | String) -> bool
+    def actor_message?(name)
+      actor_definition.messages.key?(name.to_sym)
+    end
+
+    # @rbs (Symbol | String) -> bool
+    def actor_query?(name)
+      actor_definition.queries.key?(name.to_sym)
+    end
+
+    # @rbs () -> ActorDefinition
+    def actor_definition
+      SolidObjects.registry.fetch(actor_type).definition
     end
   end
 end